apps/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.6.1
cryptography==42.0.2
python-dotenv==1.0.1
email-validator==2.1.0.post1 
# Dependências adicionais para funcionalidade completa
httpx==0.27.0
asyncio-mqtt==0.16.1
pydantic-settings==2.2.1
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.44b0
prometheus-client==0.20.0
websockets==12.0
aiofiles==23.2.1
redis==5.0.1
Pillow==10.1.0
reportlab==4.0.7
qrcode==7.4.2 